1) Took isWon() out of Level's responsibilty. Under Policy now.
2) Changed the design pattern to an MVC design pattern - With a
BlockingQueue storing commands in a very generic way. Our controller
pushes the Concrete Commands we get from our Observables to the
BlockingQueue. We get from our Observables a LinkedList<String> params -
because our commands can be 1 or 2 words (move up, display, etc). So
I've made a LinkedList<String> called 'params' in every concrete command
we have, we update this list in our controller when it removes the first
parameter of the list, and assigning it with a concrete command - then
pushes it to the queue. The queue executes the command and our params
list have the rest of the total sentence given by the user. This saves a
lot of code and is very generic.

3) Made a client/server. ATM only for 1 client each time.