3rd Milestone:
1) Added Hibernate:
a. If a user wins a level, he gets a dialog to place his user name.
Then, the game saves his stats (total steps and the finish time [in
seconds]).
b. There is a new "Highscore" feature, that shows the user a table with
each level's stats per each user name.

2) Fixed some bugs