5th Milestone: Desktop Client app
The project itself now acts as a desktop app, which represents the
client.

When a client get stuck in a game, he can request a hint for the game or
a full solution. 
A hint is 50% of the solution steps, per position.

When the client asks for a solution (or hint), the client creates a
socket connection to the server. The server contacts a web service to
check whether a solution for this level has already been saved in the
database, if not - it creates one and saves it. Afterwards the solution
is then transferred to the client via Client/Server method.