5th Milestone: The Sokoban Server
This project is the Sokoban server, the server which a client contacts
if he needs a level to be solved or given a hint.

The server, when the request is given, contacts a web service and the
web service "asks" the database whether this level's solution is already
saved in the database. Level names are unique, therefore this is
possible.

If the query from the database returns a null, then the server uses the
SokobanSolver to solve the level. It then saves the level in the
database and sends the solution to the user.

The connections is a simple socket connection that supports a thread
pool for simultaneous requests from clients

The server has a simple Admin GUI to check how many users are connected
with an option to drop the users from the table as well.